4267 경고 제거(타입이 달라서). 주항님이 만든 코드는 명시적으로 형변환, 외부에서 가져온 것은 4267 경고 무효화.

+++ BaseLayer/DelayedDirectoryChangeHandler.cpp
@@ -977,7 +977,7 @@ DWORD GetPathOffsetBasedOnFilterFlags(CDirChangeNotification * pNot, DWORD dwFil
 		//set the offset to support FILTERS_CHECK_FILE_NAME_ONLY
 		TCHAR * pSlash  = _tcsrchr(pNot->m_szFileName1, _T('\\'));
 		if( pSlash )
-			dwFileNameOffset = (++pSlash - pNot->m_szFileName1);
+			dwFileNameOffset = (DWORD)(++pSlash - pNot->m_szFileName1);
 
 		//
 		//	Because file name change notifications take place in the same directory,
